Fix client bugs
1. PostHog can only be used client-side
2. Can't nest <a> tags in the ProjectMenu
